+++ test_helpers.go
@@ -14,6 +14,11 @@ type testEvent struct {
 	Value int
 }
 
+type genericTestEvent[T any] struct {
+	Event[*testStateMachine, *testStateMachine]
+	Payload T
+}
+
 const testStateMachineID = "testStateMachine"
 
 func newTestState(name string) *testState {